refactor(hook): API_PATH 상수 사용하도록 변경
