+++ src/features/Gift/hooks/useThemes.ts
@@ -1,16 +1,16 @@
 import { useSuspenseQuery } from '@tanstack/react-query'
-import { apiGet } from '@/lib/axios'
-import type { Theme } from '../types/GiftTypes'
+import { apiGet, API_PATH } from '@/lib/axios'
+import type { Theme } from '@/types/CommonTypes'
 
 const fetchThemes = async (): Promise<Theme[]> => {
-  const res = await apiGet<Theme[]>('/themes')
+  const res = await apiGet<Theme[]>(API_PATH.THEMES)
   return res
 }
 
 export const useThemes = () => {
-  const { data } = useSuspenseQuery({
+  const { data: themes } = useSuspenseQuery({
     queryKey: ['themes'],
     queryFn: fetchThemes,
   })
-  return data
+  return themes
 }
